backend-discord/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
supabase==2.3.0
SpeechRecognition==3.10.0
pyaudio==0.2.13
numpy==1.24.3
groq==0.4.1 